Norminette and refactoring.
