+++ Makefile
@@ -6,7 +6,7 @@
 #    By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/02/13 13:55:25 by irychkov          #+#    #+#              #
-#    Updated: 2025/03/03 18:41:25 by irychkov         ###   ########.fr        #
+#    Updated: 2025/03/03 19:12:55 by irychkov         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -50,12 +50,12 @@ COMMON_SRCS =	main.c\
 				utils/libc_utils4.c\
 				utils/gnl_utils.c\
 				utils/gnl.c\
-				utils.c\
 				init.c\
 				scene.c\
 				camera.c\
 				mlx.c\
 				debug.c\
+				free.c\
 
 # Program-specific sources
 SRCS = $(COMMON_SRCS)\
