+++ src/init.c
@@ -6,41 +6,11 @@
 /*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/28 11:17:37 by irychkov          #+#    #+#             */
-/*   Updated: 2025/03/03 18:42:52 by irychkov         ###   ########.fr       */
+/*   Updated: 2025/03/03 19:05:50 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
-
-/* t_material	material(t_tuple color, double ambient, double diffuse,
-	double specular, double shininess, int has_pattern)
-{
-	t_material	material;
-
-	material.color = color;
-	material.ambient = ambient;
-	material.diffuse = diffuse;
-	material.specular = specular;
-	material.shininess = shininess;
-	material.has_pattern = has_pattern;
-	material.reflective = 0.0;
-	material.transparency = 0.0;
-	material.refractive_index = 1.0;
-	return (material);
-} */
-
-/* t_material default_material()
-{
-	t_material mat;
-	mat.color = create_color(1, 1, 1);
-	mat.ambient = 0.1;
-	mat.diffuse = 0.9;
-	mat.specular = 0.9;
-	mat.shininess = 200.0;
-	mat.has_pattern = 0;
-	mat.reflective = 0.0;
-	return (mat);
-} */
 
 t_light	init_light(t_tuple position, t_tuple color, double brightness)
 {
@@ -77,6 +47,20 @@ void	init_scene_pixels(t_scene *scene, int height, int width)
 		}
 		i++;
 	}
+}
+
+static int	validate_file_ext(t_rt *rt)
+{
+	const char	*filename;
+	const char	*dot;
+
+	filename = rt->filename;
+	dot = ft_strrchr(filename, '.');
+	if(!dot)
+		return(1);
+	if (ft_strncmp(dot, ".rt\0", 4))
+		return (1);
+	return (0);
 }
 
 void	initialize_structs(char **argv, t_rt *rt)
