+++ src/parser/parser_utils.c
@@ -3,14 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   parser_utils.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: henbuska <henbuska@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/17 10:43:18 by henbuska          #+#    #+#             */
-/*   Updated: 2025/02/28 16:54:24 by henbuska         ###   ########.fr       */
+/*   Updated: 2025/03/03 18:59:56 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
+
+int	error(char *message, int ret)
+{
+	ft_putendl_fd("Error", 2);
+	ft_putendl_fd(message, 2);
+	return (ret);
+}
 
 int	invalid_file_content(t_rt *rt)
 {
@@ -45,4 +52,42 @@ int	count_lines_in_file(int fd)
 		line = get_next_line(fd);
 	}
 	return (line_count);
+}
+
+static int	count_non_shape_elements(char **lines)
+{
+	int	count;
+	int	i;
+
+	count = 0;
+	i = 0;
+	while (lines[i])
+	{
+		if (lines[i][0] == 'A' || lines[i][0] == 'L' || lines[i][0] == 'C')
+			count++;
+		i++;
+	}
+	return (count);
+}
+
+int	init_scene_structs(char **lines, t_rt *rt)
+{
+	int	non_shape;
+
+	rt->scene = ft_calloc(1, sizeof(t_scene));
+	if (!rt->scene)
+	{
+		print_error("Failed to allocate t_scene");
+		return (1);
+	}
+	non_shape = count_non_shape_elements(lines);
+	rt->scene->shape_count = 0;
+	rt->scene->needs_render = 1;
+	rt->scene->shapes = ft_calloc(rt->element_count - non_shape, sizeof(t_shape));
+	if (!rt->scene->shapes)
+	{
+		print_error("Failed to allocate shapes array");
+		return (1);
+	}
+	return (0);
 }
